#334 on analysis page make molecule selection dependent on widget click instead of image click
