+++ src/components/analysis/molecule_image.js
@@ -1,17 +1,11 @@
 import React from "react";
 import "./analysis.css";
 import { connect } from "react-redux";
-import { selectorActions } from "../../actions";
 
 class MoleculeImage extends React.Component {
   constructor(props) {
     super(props);
   }
-
-  // fires the selectMolecule action when clicking the molecule's image
-  selectMolecule = () => {
-    this.props.selectMolecule(this.props.mol_id);
-  };
 
   render() {
     return (
@@ -19,7 +13,6 @@ class MoleculeImage extends React.Component {
         <img
           src={this.props.saved_mols[this.props.mol_id].data.img_html}
           alt="Drug"
-          onClick={this.selectMolecule}
         />
       </div>
     );
@@ -32,8 +25,4 @@ function mapStateToProps(state) {
   };
 }
 
-const actionCreators = {
-  selectMolecule: selectorActions.selectMolecule,
-};
-
-export default connect(mapStateToProps, actionCreators)(MoleculeImage);
+export default connect(mapStateToProps)(MoleculeImage);
